+++ src/holyc_lib.ml
@@ -21,6 +21,8 @@ module Ir_control_flow = Ir.Control_flow
 module Ir_block_graph = Ir.Block_graph
 module Ir_effects = Ir.Effects
 module Ir_x87_stack = Ir.X87_stack
+module Ir_integer_globals = Ir.Integer_globals
+module Ir_global_address_lowering = Ir.Global_address_lowering
 module Ir_integer_interpreter = Ir.Integer_interpreter
 module Ir_integer_program_lowering = Ir.Integer_program_lowering
 module Ir_integer_unary_folding = Ir.Integer_unary_folding
@@ -463,6 +465,7 @@ type integer_program = Driver.Integer_program.compiled
 
 let compile_integer_program = Driver.Integer_program.compile
 let integer_program_entry = Driver.Integer_program.entry
+let integer_program_globals = Driver.Integer_program.globals
 let integer_program_functions = Driver.Integer_program.functions
 let integer_program_human = Driver.Integer_program.human
 let run_integer_program = Driver.Integer_program.run